+++ src/org/jgroups/protocols/RingBufferBundler.java
@@ -134,7 +134,7 @@ public class RingBufferBundler extends BaseBundler {
         for(;;) {
             Message msg=buf[start_index];
             if(msg != null && Objects.equals(dest, msg.dest())) {
-                long size=msg.size();
+                int size=msg.size();
                 if(bytes + size > max_bundle_size)
                     break;
                 bytes+=size;
